--- client/scripts/views/components/comment_reaction_button.tsx
@@ -1,238 +1,0 @@
-/* @jsx m */
-
-import m from 'mithril';
-import mixpanel from 'mixpanel-browser';
-import { Popover } from 'construct-ui';
-import $ from 'jquery';
-
-import 'components/reaction_button.scss';
-
-import app from 'state';
-import TopicGateCheck from 'controllers/chain/ethereum/gatedTopic';
-import {
-  Proposal,
-  OffchainComment,
-  OffchainThread,
-  AnyProposal,
-  AddressInfo,
-  ChainInfo,
-} from 'models';
-import User from 'views/components/widgets/user';
-import SelectAddressModal from '../modals/select_address_modal';
-import { LoginModal } from '../modals/login_modal';
-import { CWIcon } from './component_kit/cw_icons/cw_icon';
-
-const MAX_VISIBLE_REACTING_ACCOUNTS = 10;
-
-type ReactionType = 'discussionRow' | 'threadComment';
-
-type ReactorAttrs = {
-  likes: number;
-  reactors: any;
-};
-
-type Post = OffchainThread | AnyProposal | OffchainComment<any>;
-
-type ReactionButtonAttrs = {
-  post: Post;
-  reactionType: ReactionType;
-};
-
-const getDisplayedReactorsForPopup = (reactorAttrs: ReactorAttrs) => {
-  const { reactors = [], likes = 0 } = reactorAttrs;
-
-  const slicedReactors = reactors
-    .slice(0, MAX_VISIBLE_REACTING_ACCOUNTS)
-    .map((rxn) => {
-      const {
-        Address: { address, chain },
-      } = rxn;
-
-      return m(User, {
-        user: new AddressInfo(null, address, chain, null),
-        linkify: true,
-      });
-    });
-
-  if (slicedReactors.length < likes) {
-    const diff = likes - slicedReactors.length;
-
-    slicedReactors.push(<>{`and ${diff} more`}</>);
-  }
-
-  return slicedReactors;
-};
-
-const fetchReactionsByPost = async (post: Post) => {
-  let thread_id, proposal_id, comment_id;
-
-  if (post instanceof OffchainThread) {
-    thread_id = (post as OffchainThread).id;
-  } else if (post instanceof Proposal) {
-    proposal_id = `${(post as AnyProposal).slug}_${
-      (post as AnyProposal).identifier
-    }`;
-  } else if (post instanceof OffchainComment) {
-    comment_id = (post as OffchainComment<any>).id;
-  }
-
-  const { result = [] } = await $.get(`${app.serverUrl()}/bulkReactions`, {
-    thread_id,
-    comment_id,
-    proposal_id,
-  });
-
-  return result;
-};
-
-const onReactionClick = (
-  e: MouseEvent,
-  hasReacted: boolean,
-  dislike: (userAddress: string) => void,
-  like: (chain: ChainInfo, chainId: string, userAddress: string) => void,
-  post: Post
-) => {
-  e.preventDefault();
-  e.stopPropagation();
-
-  if (!app.isLoggedIn()) {
-    app.modals.create({
-      modal: LoginModal,
-    });
-  } else if (!app.user.activeAccount) {
-    app.modals.create({
-      modal: SelectAddressModal,
-    });
-  } else {
-    const { address: userAddress, chain } = app.user.activeAccount;
-
-    // if it's a community use the app.user.activeAccount.chain.id instead of author chain
-    const chainId = app.activeChainId();
-
-    if (hasReacted) {
-      dislike(userAddress);
-    } else {
-      like(chain, chainId, userAddress);
-    }
-
-    mixpanel.track('Create Reaction ', {
-      'Step No': 1,
-      Step: 'Create Reaction',
-      'Post Name': `${post.slug}: ${post.identifier}`,
-      Scope: app.activeChainId(),
-    });
-    mixpanel.people.increment('Reaction');
-    mixpanel.people.set({
-      'Last Reaction Created': new Date().toISOString(),
-    });
-  }
-};
-
-export class ReactionButton implements m.ClassComponent<ReactionButtonAttrs> {
-  private loading: boolean;
-  private reactors: any;
-
-  oninit() {
-    this.loading = false;
-  }
-
-  view(vnode: m.VnodeDOM<ReactionButtonAttrs, this>) {
-    const { post, reactionType } = vnode.attrs;
-    const reactionCounts = app.reactionCounts.getByPost(post);
-    const { likes = 0, hasReacted } = reactionCounts || {};
-
-    // token balance check if needed
-    const isAdmin =
-      app.user.isSiteAdmin ||
-      app.user.isAdminOfEntity({ chain: app.activeChainId() });
-    let topicName = '';
-    if (post instanceof OffchainThread && post.topic && app.topics) {
-      topicName = (post as OffchainThread).topic.name;
-    } else if (post instanceof OffchainComment) {
-      // post.rootProposal has typescript typedef number but in practice seems to be a string
-      const parentThread = app.threads.getById(
-        parseInt(post.rootProposal.toString().split('_')[1], 10)
-      );
-      topicName = parentThread.topic.name;
-    }
-    this.loading =
-      vnode.state.loading ||
-      (!isAdmin && TopicGateCheck.isGatedTopic(topicName));
-
-    const activeAddress = app.user.activeAccount?.address;
-
-    const dislike = async (userAddress: string) => {
-      const reaction = (await fetchReactionsByPost(post)).find((r) => {
-        return r.Address.address === activeAddress;
-      });
-      this.loading = true;
-      app.reactionCounts
-        .delete(reaction, {
-          ...reactionCounts,
-          likes: likes - 1,
-          hasReacted: false,
-        })
-        .then(() => {
-          this.reactors = this.reactors.filter(
-            ({ Address }) => Address.address !== userAddress
-          );
-          this.loading = false;
-          m.redraw();
-        });
-    };
-
-    const like = (chain: ChainInfo, chainId: string, userAddress: string) => {
-      this.loading = true;
-      app.reactionCounts.create(userAddress, post, 'like', chainId).then(() => {
-        this.loading = false;
-        this.reactors = [
-          ...this.reactors,
-          {
-            Address: { address: userAddress, chain },
-          },
-        ];
-        m.redraw();
-      });
-    };
-
-    const reactionButtonComponent = (
-      <div
-        onmouseenter={async () => {
-          this.reactors = await fetchReactionsByPost(post);
-        }}
-        onclick={async (e) =>
-          onReactionClick(e, hasReacted, dislike, like, post)
-        }
-        class={`${
-          reactionType === 'discussionRow'
-            ? 'ReactionButton'
-            : 'ThreadReactionButton'
-        }${this.loading ? ' disabled' : ''}${hasReacted ? ' has-reacted' : ''}`}
-      >
-        <CWIcon
-          iconName={hasReacted ? 'heartFilled' : 'heartEmpty'}
-          iconSize={reactionType === 'discussionRow' ? 'small' : 'medium'}
-        />
-        <div class="reactions-count">{likes}</div>
-      </div>
-    );
-
-    return likes > 0 ? (
-      <Popover
-        interactionType="hover"
-        content={
-          <div class="reaction-button-tooltip-contents">
-            {getDisplayedReactorsForPopup({
-              likes,
-              reactors: this.reactors,
-            })}
-          </div>
-        }
-        trigger={reactionButtonComponent}
-        hoverOpenDelay={100}
-      />
-    ) : (
-      reactionButtonComponent
-    );
-  }
-}
