+++ client/scripts/views/pages/discussions/discussion_row.tsx
@@ -26,12 +26,12 @@ import {
   OffchainThreadStage,
   AddressInfo,
 } from 'models';
-import { ReactionButton } from 'views/components/reaction_button';
 import User from 'views/components/widgets/user';
 import UserGallery from 'views/components/widgets/user_gallery';
 import { DiscussionRowMenu } from './discussion_row_menu';
 import { getLastUpdated, isHot } from './helpers';
 import { CWIcon } from '../../components/component_kit/cw_icons/cw_icon';
+import { DiscussionRowReactionButton } from '../../components/reaction_button/discussion_row_reaction_button';
 
 type DiscussionRowAttrs = {
   onSelect?: any;
@@ -39,7 +39,7 @@ type DiscussionRowAttrs = {
 };
 
 export class DiscussionRow implements m.ClassComponent<DiscussionRowAttrs> {
-  view(vnode) {
+  view(vnode: m.VnodeDOM<DiscussionRowAttrs, this>) {
     const { proposal } = vnode.attrs;
 
     const discussionLink = getProposalUrlPath(
@@ -68,7 +68,7 @@ export class DiscussionRow implements m.ClassComponent<DiscussionRowAttrs> {
             <CWIcon iconName="pin" iconSize="small" />
           </div>
         ) : (
-          <ReactionButton post={proposal} reactionType="discussionRow" />
+          <DiscussionRowReactionButton thread={proposal} />
         )}
         <div class="title-container">
           <div class="row-header">{proposal.title}</div>
