+++ client/styles/components/reaction_button/comment_reaction_button.scss
@@ -1,41 +1,6 @@
 @import 'shared';
 
-.ReactionButton {
-  align-items: center;
-  border-radius: 12px;
-  border: 1px solid $neutral-300;
-  color: $neutral-500;
-  cursor: pointer;
-  display: flex;
-  flex-direction: column;
-  font-weight: 500;
-  height: 48px;
-  justify-content: center;
-  margin-right: 24px;
-  min-width: 48px;
-
-  &.has-reacted {
-    border-color: $primary-500;
-    color: $primary-500;
-  }
-
-  &.disabled {
-    border-color: $neutral-50;
-    color: $neutral-500;
-    pointer-events: none;
-  }
-
-  svg {
-    margin: 4px auto 2px;
-  }
-
-  .reactions-count {
-    font-size: 14px;
-    margin: 2px auto 4px;
-  }
-}
-
-.ThreadReactionButton {
+.CommentReactionButton {
   align-items: center;
   border: 1px solid $neutral-500;
   border-radius: 20px;
